% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actphys_sedent_decode.R
\docType{data}
\name{actphys_sedent_decode}
\alias{actphys_sedent_decode}
\title{actphys_sedent_decode}
\format{A data frame with 4725 rows and 93 variables:
\describe{
  \item{ POPULATION }{  character }
  \item{ NOIND }{  numeric }
  \item{ transport_personnel }{  character }
  \item{ transport_ecole }{  character }
  \item{ tv_score }{  logical }
  \item{ tv_duree }{  numeric }
  \item{ jvideo_score }{  logical }
  \item{ jvideo_duree }{  numeric }
  \item{ ordi_score }{  logical }
  \item{ ordi_duree }{  numeric }
  \item{ jeux_pleinair_nbjours }{  numeric }
  \item{ enfant_actif }{  character }
  \item{ eps_ouinon }{  character }
  \item{ eps_nbjours }{  numeric }
  \item{ eps_duree }{  character }
  \item{ club_nbjours }{  numeric }
  \item{ club_nb }{  character }
  \item{ activite_musculation_nbjours }{  numeric }
  \item{ activite_moderee_nbjours }{  numeric }
  \item{ activite_intense_nbjours }{  numeric }
  \item{ escalier_score }{  logical }
  \item{ emploi_4sem }{  character }
  \item{ travail_freq_semaine }{  logical }
  \item{ travail_duree }{  logical }
  \item{ travail_score }{  logical }
  \item{ transport_voiture_score }{  logical }
  \item{ transport_public_score }{  logical }
  \item{ transport_velo_score }{  logical }
  \item{ transport_pied_score }{  logical }
  \item{ transport_score }{  logical }
  \item{ activite_balayer_score }{  logical }
  \item{ activite_nettoyer_score }{  logical }
  \item{ activite_aspirateur_score }{  logical }
  \item{ activite_repasser_score }{  logical }
  \item{ activite_vaisselle_score }{  logical }
  \item{ activite_lessive_score }{  logical }
  \item{ activite_jardiner_score }{  logical }
  \item{ activite_tondre_score }{  logical }
  \item{ activite_arroser_score }{  logical }
  \item{ activite_becher_score }{  logical }
  \item{ activite_bricoler_score }{  logical }
  \item{ activite_marche_score }{  logical }
  \item{ activite_piscine_score }{  logical }
  \item{ activite_velo_score }{  logical }
  \item{ activite_roller_score }{  logical }
  \item{ activite_danse_salon_score }{  logical }
  \item{ activite_bowling_score }{  logical }
  \item{ activite_billard_score }{  logical }
  \item{ activite_peche_score }{  logical }
  \item{ activite_chasse_score }{  logical }
  \item{ activite_loisirs_assis_score }{  logical }
  \item{ activite_loisirs_debout_score }{  logical }
  \item{ activite_randonnee_score }{  logical }
  \item{ activite_jogging_score }{  logical }
  \item{ activite_natation_score }{  logical }
  \item{ activite_yoga_score }{  logical }
  \item{ activite_rameur_score }{  logical }
  \item{ activite_aerobic_score }{  logical }
  \item{ activite_danse_moderne_score }{  logical }
  \item{ activite_cyclisme_score }{  logical }
  \item{ activite_tennis_score }{  logical }
  \item{ activite_boxe_score }{  logical }
  \item{ activite_escrime_score }{  logical }
  \item{ activite_haltere_score }{  logical }
  \item{ activite_athletisme_score }{  logical }
  \item{ activite_ski_score }{  logical }
  \item{ activite_voile_score }{  logical }
  \item{ activite_canoe_score }{  logical }
  \item{ activite_equitation_score }{  logical }
  \item{ activite_golf_score }{  logical }
  \item{ activite_tir_score }{  logical }
  \item{ activite_moto_score }{  logical }
  \item{ activite_musique_score }{  logical }
  \item{ activite_ping_pong_score }{  logical }
  \item{ activite_gym_score }{  logical }
  \item{ activite_escalade_score }{  logical }
  \item{ activite_alpinisme_score }{  logical }
  \item{ activite_foot_score }{  logical }
  \item{ activite_rugby_score }{  logical }
  \item{ activite_basket_score }{  logical }
  \item{ activite_hand_score }{  logical }
  \item{ activite_volley_score }{  logical }
  \item{ activite_hockey_score }{  logical }
  \item{ activite_domloissport_duree }{  logical }
  \item{ activite_domloissport_score }{  logical }
  \item{ activite_moderee_score }{  logical }
  \item{ activite_intense_score }{  logical }
  \item{ activite_total_duree }{  logical }
  \item{ activite_total_score }{  logical }
  \item{ nap }{  character }
  \item{ profil_activite }{  character }
  \item{ sedentarite }{  character }
  \item{ sedentarite_duree }{  logical }
}}
\source{
Inca3
}
\usage{
actphys_sedent_decode
}
\description{
Data inca3.
}
\keyword{datasets}
