% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apports_nut_alim_decode.R
\docType{data}
\name{apports_nut_alim_decode}
\alias{apports_nut_alim_decode}
\title{apports_nut_alim_decode}
\format{

}
\source{
Inca3
}
\usage{
apports_nut_alim_decode
}
\description{
Data inca3.
}
\keyword{datasets}
