% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conso_ca_indiv_decode.R
\docType{data}
\name{conso_ca_indiv_decode}
\alias{conso_ca_indiv_decode}
\title{conso_ca_indiv_decode}
\format{A data frame with 5855 rows and 18 variables:
\describe{
  \item{ POPULATION }{  character }
  \item{ NOIND }{  numeric }
  \item{ periode_reference }{  character }
  \item{ conso_ca }{  character }
  \item{ conso_ca_regl }{  character }
  \item{ conso_ca_nb }{  numeric }
  \item{ conso_ca_regl_nb }{  numeric }
  \item{ indic_ca }{  numeric }
  \item{ indic_medic }{  numeric }
  \item{ indic_ni }{  numeric }
  \item{ saison_particuliere }{  character }
  \item{ saison_automne }{  character }
  \item{ saison_hiver }{  character }
  \item{ saison_printemps }{  character }
  \item{ saison_ete }{  character }
  \item{ saison_autre }{  character }
  \item{ saison_autre_libelle }{  character }
  \item{ saison_nsp }{  character }
}}
\source{
Inca3
}
\usage{
conso_ca_indiv_decode
}
\description{
Data inca3.
}
\keyword{datasets}
