% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conso_ca_prod.R
\docType{data}
\name{conso_ca_prod}
\alias{conso_ca_prod}
\title{conso_ca_prod}
\format{A data frame with 2380 rows and 12 variables:
\describe{
  \item{ POPULATION }{  character }
  \item{ NOIND }{  numeric }
  \item{ periode_reference }{  numeric }
  \item{ num_ligne_CA }{  numeric }
  \item{ num_prod }{  numeric }
  \item{ type_prod }{  numeric }
  \item{ classif_reg_prod }{  character }
  \item{ classif_prod }{  numeric }
  \item{ pres_prod }{  numeric }
  \item{ nb_unit_prod }{  numeric }
  \item{ mode_conso_prod }{  numeric }
  \item{ nb_jours_an }{  numeric }
}}
\source{
Inca3
}
\usage{
conso_ca_prod
}
\description{
Data inca3.
}
\keyword{datasets}
