% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conso_gpe_inca3.R
\docType{data}
\name{conso_gpe_inca3}
\alias{conso_gpe_inca3}
\title{conso_gpe_inca3}
\format{A data frame with 4114 rows and 46 variables:
\describe{
  \item{ POPULATION }{  character }
  \item{ NOIND }{  numeric }
  \item{ conso_gpe1 }{  numeric }
  \item{ conso_gpe2 }{  numeric }
  \item{ conso_gpe3 }{  numeric }
  \item{ conso_gpe4 }{  numeric }
  \item{ conso_gpe5 }{  numeric }
  \item{ conso_gpe6 }{  numeric }
  \item{ conso_gpe7 }{  numeric }
  \item{ conso_gpe8 }{  numeric }
  \item{ conso_gpe9 }{  numeric }
  \item{ conso_gpe10 }{  numeric }
  \item{ conso_gpe11 }{  numeric }
  \item{ conso_gpe12 }{  numeric }
  \item{ conso_gpe13 }{  numeric }
  \item{ conso_gpe14 }{  numeric }
  \item{ conso_gpe15 }{  numeric }
  \item{ conso_gpe16 }{  numeric }
  \item{ conso_gpe17 }{  numeric }
  \item{ conso_gpe18 }{  numeric }
  \item{ conso_gpe19 }{  numeric }
  \item{ conso_gpe20 }{  numeric }
  \item{ conso_gpe21 }{  numeric }
  \item{ conso_gpe22 }{  numeric }
  \item{ conso_gpe23 }{  numeric }
  \item{ conso_gpe24 }{  numeric }
  \item{ conso_gpe25 }{  numeric }
  \item{ conso_gpe26 }{  numeric }
  \item{ conso_gpe27 }{  numeric }
  \item{ conso_gpe28 }{  numeric }
  \item{ conso_gpe29 }{  numeric }
  \item{ conso_gpe30 }{  numeric }
  \item{ conso_gpe31 }{  numeric }
  \item{ conso_gpe32 }{  numeric }
  \item{ conso_gpe33 }{  numeric }
  \item{ conso_gpe34 }{  numeric }
  \item{ conso_gpe35 }{  numeric }
  \item{ conso_gpe36 }{  numeric }
  \item{ conso_gpe37 }{  numeric }
  \item{ conso_gpe38 }{  numeric }
  \item{ conso_gpe39 }{  numeric }
  \item{ conso_gpe40 }{  numeric }
  \item{ conso_gpe41 }{  numeric }
  \item{ conso_gpe42 }{  numeric }
  \item{ conso_gpe43 }{  numeric }
  \item{ conso_gpe44 }{  numeric }
}}
\source{
Inca3
}
\usage{
conso_gpe_inca3
}
\description{
Data inca3.
}
\keyword{datasets}
