% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_indiv.R
\docType{data}
\name{description_indiv}
\alias{description_indiv}
\title{description_indiv}
\format{
A data frame with 5855 rows and 185 variables:
\describe{
  \item{ NOMEN }{  numeric }
  \item{ NOIND }{  numeric }
  \item{ ech }{  numeric }
  \item{ enf_allaite }{  numeric }
  \item{ pop1 }{  numeric }
  \item{ pop2 }{  numeric }
  \item{ pop3 }{  numeric }
  \item{ pond_indiv_adu_pop1 }{  numeric }
  \item{ pond_indiv_enf_pop1 }{  numeric }
  \item{ pond_indiv_adu_pop2 }{  numeric }
  \item{ pond_indiv_enf_pop2 }{  numeric }
  \item{ pond_indiv_adu_pop3 }{  numeric }
  \item{ pond_indiv_enf_pop3 }{  numeric }
  \item{ pond_men_pop1 }{  numeric }
  \item{ pond_men_pop2 }{  numeric }
  \item{ zae }{  character }
  \item{ strate }{  numeric }
  \item{ fpc1 }{  numeric }
  \item{ fpc2 }{  numeric }
  \item{ fpc3 }{  numeric }
  \item{ saison_pop1 }{  numeric }
  \item{ saison_pop2 }{  numeric }
  \item{ saison_pop3 }{  numeric }
  \item{ region_adm_12cl }{  numeric }
  \item{ region_inca3 }{  numeric }
  \item{ agglo_5cl }{  numeric }
  \item{ sex_PS }{  numeric }
  \item{ tage_PS }{  numeric }
  \item{ tage_PS_mois }{  numeric }
  \item{ lien_rep_enf }{  numeric }
  \item{ diplome_interv }{  numeric }
  \item{ etude_4cl_interv }{  numeric }
  \item{ situ_prof_5cl_interv }{  numeric }
  \item{ atrav_interv }{  numeric }
  \item{ trav_nuit_interv }{  numeric }
  \item{ trav_nuit_2cl_interv }{  numeric }
  \item{ PCS_8cl_interv }{  numeric }
  \item{ PCS_4cl_interv }{  numeric }
  \item{ tps_travail_interv }{  numeric }
  \item{ vacances_interv }{  numeric }
  \item{ interv_PR }{  numeric }
  \item{ sex_PR }{  numeric }
  \item{ tage_PR }{  numeric }
  \item{ lien_interv_PR }{  numeric }
  \item{ lien_PS_PR }{  numeric }
  \item{ diplome_PR }{  numeric }
  \item{ etude_4cl_PR }{  numeric }
  \item{ atrav_PR }{  numeric }
  \item{ PCS_8cl_PR }{  numeric }
  \item{ PCS_4cl_PR }{  numeric }
  \item{ tps_travail_PR }{  numeric }
  \item{ stat_log_2cl }{  numeric }
  \item{ soins }{  numeric }
  \item{ situ_fin_3cl }{  numeric }
  \item{ revenu }{  numeric }
  \item{ RUC_4cl }{  numeric }
  \item{ nbpers }{  numeric }
  \item{ nbadu }{  numeric }
  \item{ nbenf }{  numeric }
  \item{ situ_alim_statut }{  numeric }
  \item{ IA_statut }{  numeric }
  \item{ IA_score }{  numeric }
  \item{ taille_m }{  numeric }
  \item{ taille_d }{  numeric }
  \item{ taille }{  numeric }
  \item{ poids_m }{  numeric }
  \item{ poids_d }{  numeric }
  \item{ poids }{  numeric }
  \item{ imc }{  numeric }
  \item{ statnut }{  numeric }
  \item{ maladie_allergie_alim }{  numeric }
  \item{ intoall_confirm_med }{  numeric }
  \item{ regime_vegetarien }{  numeric }
  \item{ regime_allergie }{  numeric }
  \item{ regime_maigrir_med }{  numeric }
  \item{ regime_maigrir_choix }{  numeric }
  \item{ regime_autre_med }{  numeric }
  \item{ regime_raisonmed_libelle }{  character }
  \item{ regime_poidsstable }{  numeric }
  \item{ regime_forme }{  numeric }
  \item{ regime_autreraison }{  numeric }
  \item{ regime_non }{  numeric }
  \item{ veget_viande }{  numeric }
  \item{ veget_prodmer }{  numeric }
  \item{ veget_prodlait }{  numeric }
  \item{ veget_oeuf }{  numeric }
  \item{ veget_miel }{  numeric }
  \item{ veget_autre_alim }{  numeric }
  \item{ veget_autre_alim_libelle }{  logical }
  \item{ allergie_laitvache }{  numeric }
  \item{ allergie_prepainfsoja }{  numeric }
  \item{ allergie_prepainfamande }{  numeric }
  \item{ allergie_gluten }{  numeric }
  \item{ allergie_farineble }{  numeric }
  \item{ allergie_lupin }{  numeric }
  \item{ allergie_arachide }{  numeric }
  \item{ allergie_fruitcoque }{  numeric }
  \item{ allergie_fruitcoque_libelle }{  character }
  \item{ allergie_oeuf }{  numeric }
  \item{ allergie_poisson }{  numeric }
  \item{ allergie_crustace }{  numeric }
  \item{ allergie_mollusque }{  numeric }
  \item{ allergie_soja }{  numeric }
  \item{ allergie_sesame }{  numeric }
  \item{ allergie_moutarde }{  numeric }
  \item{ allergie_sulfite }{  numeric }
  \item{ allergie_celeri }{  numeric }
  \item{ allergie_autres_fruitleg }{  numeric }
  \item{ allergie_autres_fl_libelle }{  character }
  \item{ allergie_autresalim }{  numeric }
  \item{ allergie_autresalim_libelle }{  character }
  \item{ allergie_nondetermine }{  numeric }
  \item{ allergie_fruits }{  numeric }
  \item{ allergie_legumes }{  numeric }
  \item{ regime_passe }{  numeric }
  \item{ regime_nb_2dernann }{  numeric }
  \item{ regime_nb_anter2dernann }{  numeric }
  \item{ regime_type }{  numeric }
  \item{ regime_type_libelle }{  character }
  \item{ regime_duree_sem }{  numeric }
  \item{ regime_duree_mois }{  numeric }
  \item{ regime_duree_nsp }{  numeric }
  \item{ poids_anndern }{  numeric }
  \item{ poids_anndern_nsp }{  numeric }
  \item{ poids_modif }{  numeric }
  \item{ poids_modifalim }{  numeric }
  \item{ poids_plusAP }{  numeric }
  \item{ poids_medicaments }{  numeric }
  \item{ poids_substituts }{  numeric }
  \item{ poids_chirurgie }{  numeric }
  \item{ poids_modifalim_laityaourt }{  numeric }
  \item{ poids_modifalim_fromage }{  numeric }
  \item{ poids_modifalim_mg }{  numeric }
  \item{ poids_modifalim_fruit }{  numeric }
  \item{ poids_modifalim_legume }{  numeric }
  \item{ poids_modifalim_pdtfeculent }{  numeric }
  \item{ poids_modifalim_pizza }{  numeric }
  \item{ poids_modifalim_pain }{  numeric }
  \item{ poids_modifalim_vrouge }{  numeric }
  \item{ poids_modifalim_volaille }{  numeric }
  \item{ poids_modifalim_oeuf }{  numeric }
  \item{ poids_modifalim_gateau }{  numeric }
  \item{ poids_modifalim_edulcorant }{  numeric }
  \item{ poids_modifalim_pdtsalleges }{  numeric }
  \item{ poids_modifalim_BS }{  numeric }
  \item{ poids_modifalim_eau }{  numeric }
  \item{ poids_modifalim_autre }{  numeric }
  \item{ poids_modifalim_autre_libelle }{  character }
  \item{ poids_perception }{  numeric }
  \item{ poidsmax }{  numeric }
  \item{ poidsmax_nsp }{  numeric }
  \item{ age_poidsmax }{  numeric }
  \item{ age_poidsmax_nsp }{  numeric }
  \item{ poidsmin }{  numeric }
  \item{ poidsmin_nsp }{  numeric }
  \item{ age_poidsmin }{  numeric }
  \item{ age_poidsmin_nsp }{  numeric }
  \item{ nb_prise_10kg }{  numeric }
  \item{ menopause }{  numeric }
  \item{ enceinte }{  numeric }
  \item{ enceinte_nbmois }{  numeric }
  \item{ allaite }{  numeric }
  \item{ allaite_nbsem }{  numeric }
  \item{ enceinte_12dermois }{  numeric }
  \item{ fume }{  numeric }
  \item{ nb_cigarettes_jour }{  numeric }
  \item{ nb_cigarettes_sem }{  numeric }
  \item{ nb_cigarettes_nsp }{  numeric }
  \item{ nb_cigares_jour }{  numeric }
  \item{ nb_cigares_sem }{  numeric }
  \item{ nb_cigares_nsp }{  numeric }
  \item{ nb_pipes_jour }{  numeric }
  \item{ nb_pipes_sem }{  numeric }
  \item{ nb_pipes_nsp }{  numeric }
  \item{ fume_age_debut }{  numeric }
  \item{ fume_age_debut_nsp }{  numeric }
  \item{ fume_age_arret }{  numeric }
  \item{ fume_age_arret_nsp }{  numeric }
  \item{ bmr_kcal }{  numeric }
  \item{ sousest0 }{  numeric }
  \item{ surest0 }{  numeric }
  \item{ sousest1 }{  numeric }
  \item{ sousest3 }{  numeric }
  \item{ sousext }{  numeric }
  \item{ surext }{  numeric }
}
}
\source{
Inca3
}
\usage{
description_indiv
}
\description{
Data inca3.
}
\keyword{datasets}
