% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_indiv_decode.R
\docType{data}
\name{description_indiv_decode}
\alias{description_indiv_decode}
\title{description_indiv_decode}
\format{A data frame with 5855 rows and 185 variables:
\describe{
  \item{ NOMEN }{  numeric }
  \item{ NOIND }{  numeric }
  \item{ ech }{  character }
  \item{ enf_allaite }{  character }
  \item{ pop1 }{  character }
  \item{ pop2 }{  character }
  \item{ pop3 }{  character }
  \item{ pond_indiv_adu_pop1 }{  numeric }
  \item{ pond_indiv_enf_pop1 }{  numeric }
  \item{ pond_indiv_adu_pop2 }{  numeric }
  \item{ pond_indiv_enf_pop2 }{  numeric }
  \item{ pond_indiv_adu_pop3 }{  numeric }
  \item{ pond_indiv_enf_pop3 }{  numeric }
  \item{ pond_men_pop1 }{  numeric }
  \item{ pond_men_pop2 }{  numeric }
  \item{ zae }{  character }
  \item{ strate }{  character }
  \item{ fpc1 }{  numeric }
  \item{ fpc2 }{  numeric }
  \item{ fpc3 }{  numeric }
  \item{ saison_pop1 }{  character }
  \item{ saison_pop2 }{  character }
  \item{ saison_pop3 }{  character }
  \item{ region_adm_12cl }{  character }
  \item{ region_inca3 }{  character }
  \item{ agglo_5cl }{  character }
  \item{ sex_PS }{  character }
  \item{ tage_PS }{  character }
  \item{ tage_PS_mois }{  character }
  \item{ lien_rep_enf }{  character }
  \item{ diplome_interv }{  character }
  \item{ etude_4cl_interv }{  character }
  \item{ situ_prof_5cl_interv }{  character }
  \item{ atrav_interv }{  character }
  \item{ trav_nuit_interv }{  character }
  \item{ trav_nuit_2cl_interv }{  character }
  \item{ PCS_8cl_interv }{  character }
  \item{ PCS_4cl_interv }{  character }
  \item{ tps_travail_interv }{  character }
  \item{ vacances_interv }{  character }
  \item{ interv_PR }{  character }
  \item{ sex_PR }{  character }
  \item{ tage_PR }{  character }
  \item{ lien_interv_PR }{  character }
  \item{ lien_PS_PR }{  character }
  \item{ diplome_PR }{  character }
  \item{ etude_4cl_PR }{  character }
  \item{ atrav_PR }{  character }
  \item{ PCS_8cl_PR }{  character }
  \item{ PCS_4cl_PR }{  character }
  \item{ tps_travail_PR }{  character }
  \item{ stat_log_2cl }{  character }
  \item{ soins }{  character }
  \item{ situ_fin_3cl }{  character }
  \item{ revenu }{  character }
  \item{ RUC_4cl }{  character }
  \item{ nbpers }{  numeric }
  \item{ nbadu }{  numeric }
  \item{ nbenf }{  numeric }
  \item{ situ_alim_statut }{  character }
  \item{ IA_statut }{  numeric }
  \item{ IA_score }{  numeric }
  \item{ taille_m }{  numeric }
  \item{ taille_d }{  numeric }
  \item{ taille }{  numeric }
  \item{ poids_m }{  numeric }
  \item{ poids_d }{  numeric }
  \item{ poids }{  numeric }
  \item{ imc }{  numeric }
  \item{ statnut }{  character }
  \item{ maladie_allergie_alim }{  character }
  \item{ intoall_confirm_med }{  character }
  \item{ regime_vegetarien }{  character }
  \item{ regime_allergie }{  character }
  \item{ regime_maigrir_med }{  character }
  \item{ regime_maigrir_choix }{  character }
  \item{ regime_autre_med }{  character }
  \item{ regime_raisonmed_libelle }{  character }
  \item{ regime_poidsstable }{  character }
  \item{ regime_forme }{  character }
  \item{ regime_autreraison }{  character }
  \item{ regime_non }{  character }
  \item{ veget_viande }{  character }
  \item{ veget_prodmer }{  character }
  \item{ veget_prodlait }{  character }
  \item{ veget_oeuf }{  character }
  \item{ veget_miel }{  character }
  \item{ veget_autre_alim }{  character }
  \item{ veget_autre_alim_libelle }{  logical }
  \item{ allergie_laitvache }{  character }
  \item{ allergie_prepainfsoja }{  character }
  \item{ allergie_prepainfamande }{  character }
  \item{ allergie_gluten }{  character }
  \item{ allergie_farineble }{  character }
  \item{ allergie_lupin }{  character }
  \item{ allergie_arachide }{  character }
  \item{ allergie_fruitcoque }{  character }
  \item{ allergie_fruitcoque_libelle }{  character }
  \item{ allergie_oeuf }{  character }
  \item{ allergie_poisson }{  character }
  \item{ allergie_crustace }{  character }
  \item{ allergie_mollusque }{  character }
  \item{ allergie_soja }{  character }
  \item{ allergie_sesame }{  character }
  \item{ allergie_moutarde }{  character }
  \item{ allergie_sulfite }{  character }
  \item{ allergie_celeri }{  character }
  \item{ allergie_autres_fruitleg }{  character }
  \item{ allergie_autres_fl_libelle }{  character }
  \item{ allergie_autresalim }{  character }
  \item{ allergie_autresalim_libelle }{  character }
  \item{ allergie_nondetermine }{  character }
  \item{ allergie_fruits }{  character }
  \item{ allergie_legumes }{  character }
  \item{ regime_passe }{  character }
  \item{ regime_nb_2dernann }{  character }
  \item{ regime_nb_anter2dernann }{  character }
  \item{ regime_type }{  character }
  \item{ regime_type_libelle }{  character }
  \item{ regime_duree_sem }{  numeric }
  \item{ regime_duree_mois }{  numeric }
  \item{ regime_duree_nsp }{  character }
  \item{ poids_anndern }{  numeric }
  \item{ poids_anndern_nsp }{  character }
  \item{ poids_modif }{  character }
  \item{ poids_modifalim }{  character }
  \item{ poids_plusAP }{  character }
  \item{ poids_medicaments }{  character }
  \item{ poids_substituts }{  character }
  \item{ poids_chirurgie }{  character }
  \item{ poids_modifalim_laityaourt }{  character }
  \item{ poids_modifalim_fromage }{  character }
  \item{ poids_modifalim_mg }{  character }
  \item{ poids_modifalim_fruit }{  character }
  \item{ poids_modifalim_legume }{  character }
  \item{ poids_modifalim_pdtfeculent }{  character }
  \item{ poids_modifalim_pizza }{  character }
  \item{ poids_modifalim_pain }{  character }
  \item{ poids_modifalim_vrouge }{  character }
  \item{ poids_modifalim_volaille }{  character }
  \item{ poids_modifalim_oeuf }{  character }
  \item{ poids_modifalim_gateau }{  character }
  \item{ poids_modifalim_edulcorant }{  character }
  \item{ poids_modifalim_pdtsalleges }{  character }
  \item{ poids_modifalim_BS }{  character }
  \item{ poids_modifalim_eau }{  character }
  \item{ poids_modifalim_autre }{  character }
  \item{ poids_modifalim_autre_libelle }{  character }
  \item{ poids_perception }{  character }
  \item{ poidsmax }{  numeric }
  \item{ poidsmax_nsp }{  character }
  \item{ age_poidsmax }{  numeric }
  \item{ age_poidsmax_nsp }{  character }
  \item{ poidsmin }{  numeric }
  \item{ poidsmin_nsp }{  character }
  \item{ age_poidsmin }{  numeric }
  \item{ age_poidsmin_nsp }{  character }
  \item{ nb_prise_10kg }{  character }
  \item{ menopause }{  character }
  \item{ enceinte }{  character }
  \item{ enceinte_nbmois }{  numeric }
  \item{ allaite }{  character }
  \item{ allaite_nbsem }{  numeric }
  \item{ enceinte_12dermois }{  character }
  \item{ fume }{  character }
  \item{ nb_cigarettes_jour }{  numeric }
  \item{ nb_cigarettes_sem }{  numeric }
  \item{ nb_cigarettes_nsp }{  numeric }
  \item{ nb_cigares_jour }{  numeric }
  \item{ nb_cigares_sem }{  numeric }
  \item{ nb_cigares_nsp }{  numeric }
  \item{ nb_pipes_jour }{  numeric }
  \item{ nb_pipes_sem }{  numeric }
  \item{ nb_pipes_nsp }{  numeric }
  \item{ fume_age_debut }{  numeric }
  \item{ fume_age_debut_nsp }{  numeric }
  \item{ fume_age_arret }{  numeric }
  \item{ fume_age_arret_nsp }{  numeric }
  \item{ bmr_kcal }{  numeric }
  \item{ sousest0 }{  character }
  \item{ surest0 }{  character }
  \item{ sousest1 }{  character }
  \item{ sousest3 }{  character }
  \item{ sousext }{  character }
  \item{ surext }{  character }
}}
\source{
Inca3
}
\usage{
description_indiv_decode
}
\description{
Data inca3.
}
\keyword{datasets}
