% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpq_decode.R
\docType{data}
\name{fpq_decode}
\alias{fpq_decode}
\title{fpq_decode}
\format{
A data frame with 4339 rows and 384 variables:
\describe{
  \item{ POPULATION }{  character }
  \item{ NOIND }{  numeric }
  \item{ PC_pain_ON }{  character }
  \item{ PC_pain_freq_M }{  numeric }
  \item{ PC_paincomplet_ON }{  character }
  \item{ PC_paincomplet_freq_M }{  numeric }
  \item{ PC_cereales_pdj_complet_ON }{  character }
  \item{ PC_cereales_pdj_complet_freq_M }{  numeric }
  \item{ PC_cereales_pdj_autre_ON }{  character }
  \item{ PC_cereales_pdj_autre_freq_M }{  numeric }
  \item{ RP_riz_tot_ON }{  character }
  \item{ RP_riz_tot_freq_M }{  numeric }
  \item{ RP_riz_complet_ON }{  character }
  \item{ RP_riz_complet_freq_M }{  numeric }
  \item{ RP_pdt_tot_ON }{  character }
  \item{ RP_pdt_tot_freq_M }{  numeric }
  \item{ RP_pdt_frites_ON }{  character }
  \item{ RP_pdt_frites_freq_M }{  numeric }
  \item{ RP_legumeracine_ON }{  character }
  \item{ RP_legumeracine_freq_M }{  numeric }
  \item{ RP_tubercule_ON }{  character }
  \item{ RP_tubercule_freq_M }{  numeric }
  \item{ LEG_tot_ON }{  character }
  \item{ LEG_tot_freq_M }{  numeric }
  \item{ LEG_salade_ON }{  character }
  \item{ LEG_salade_freq_M }{  numeric }
  \item{ LEG_bettera_tot_ON }{  character }
  \item{ LEG_bettera_tot_freq_M }{  numeric }
  \item{ LEG_bettera_carot_ON }{  character }
  \item{ LEG_bettera_carot_freq_M }{  numeric }
  \item{ LEG_concombre_ON }{  character }
  \item{ LEG_concombre_freq_M }{  numeric }
  \item{ LEG_concombre_printété_ON }{  character }
  \item{ LEG_concombre_printété_freq_M }{  numeric }
  \item{ LEG_concombre_automhiv_ON }{  character }
  \item{ LEG_concombre_automhiv_freq_M }{  numeric }
  \item{ LEG_tomate_tot_ON }{  character }
  \item{ LEG_tomate_tot_freq_M }{  numeric }
  \item{ LEG_tomate_tot_printété_ON }{  character }
  \item{ LEG_tomate_tot_printété_freq_M }{  numeric }
  \item{ LEG_tomate_tot_automhiv_ON }{  character }
  \item{ LEG_tomate_tot_automhiv_freq_M }{  numeric }
  \item{ LEG_tomate_seul_ON }{  character }
  \item{ LEG_tomate_seul_freq_M }{  numeric }
  \item{ LEG_tomate_seul_printété_ON }{  character }
  \item{ LEG_tomate_seul_printété_freq_M }{  numeric }
  \item{ LEG_tomate_seul_automhiv_ON }{  character }
  \item{ LEG_tomate_seul_automhiv_freq_M }{  numeric }
  \item{ LEG_chou_ON }{  character }
  \item{ LEG_chou_freq_M }{  numeric }
  \item{ LEG_artichaut_ON }{  character }
  \item{ LEG_artichaut_freq_M }{  numeric }
  \item{ LEG_champignon_ON }{  character }
  \item{ LEG_champignon_freq_M }{  numeric }
  \item{ LEG_oignon_ON }{  character }
  \item{ LEG_oignon_freq_M }{  numeric }
  \item{ LEG_legsecs_ON }{  character }
  \item{ LEG_legsecs_freq_M }{  numeric }
  \item{ LEG_soupe_ON }{  character }
  \item{ LEG_soupe_freq_M }{  numeric }
  \item{ LEG_soupe_automhiv_ON }{  character }
  \item{ LEG_soupe_automhiv_freq_M }{  numeric }
  \item{ LEG_soupe_printété_ON }{  character }
  \item{ LEG_soupe_printété_freq_M }{  numeric }
  \item{ LEG_saucetomate_ON }{  character }
  \item{ LEG_saucetomate_freq_M }{  numeric }
  \item{ VC_tot_ON }{  character }
  \item{ VC_tot_freq_M }{  numeric }
  \item{ VC_boeuf_ON }{  character }
  \item{ VC_boeuf_freq_M }{  numeric }
  \item{ VC_agneau_ON }{  character }
  \item{ VC_agneau_freq_M }{  numeric }
  \item{ VC_poulet_ON }{  character }
  \item{ VC_poulet_freq_M }{  numeric }
  \item{ VC_lapin_ON }{  character }
  \item{ VC_lapin_freq_M }{  numeric }
  \item{ VC_porc_ON }{  character }
  \item{ VC_porc_freq_M }{  numeric }
  \item{ VC_saucisson_ON }{  character }
  \item{ VC_saucisson_freq_M }{  numeric }
  \item{ VC_pate_foie_ON }{  character }
  \item{ VC_pate_foie_freq_M }{  numeric }
  \item{ VC_pate_autre_ON }{  character }
  \item{ VC_pate_autre_freq_M }{  numeric }
  \item{ VC_foie_ON }{  character }
  \item{ VC_foie_freq_M }{  numeric }
  \item{ PPM_tot_ON }{  character }
  \item{ PPM_tot_freq_M }{  numeric }
  \item{ PPM_poissgras_tot_ON }{  character }
  \item{ PPM_poissgras_tot_freq_M }{  numeric }
  \item{ PPM_poissgras_thon_ON }{  character }
  \item{ PPM_poissgras_thon_freq_M }{  numeric }
  \item{ PPM_poissonautre_ON }{  character }
  \item{ PPM_poissonautre_freq_M }{  numeric }
  \item{ PPM_surimi_ON }{  character }
  \item{ PPM_surimi_freq_M }{  numeric }
  \item{ PPM_crustace_ON }{  character }
  \item{ PPM_crustace_freq_M }{  numeric }
  \item{ PPM_mollusque_ON }{  character }
  \item{ PPM_mollusque_freq_M }{  numeric }
  \item{ OE_oeuf_ON }{  character }
  \item{ OE_oeuf_freq_M }{  numeric }
  \item{ OE_quiche_ON }{  character }
  \item{ OE_quiche_freq_M }{  numeric }
  \item{ SOJ_soja_ON }{  character }
  \item{ SOJ_soja_freq_M }{  numeric }
  \item{ PL_lait_ON }{  character }
  \item{ PL_lait_freq_M }{  numeric }
  \item{ PL_fromage_tot_ON }{  character }
  \item{ PL_fromage_tot_freq_M }{  numeric }
  \item{ PL_fromage_laitcru_ON }{  character }
  \item{ PL_fromage_laitcru_freq_M }{  numeric }
  \item{ PL_yaourtentremet_ON }{  character }
  \item{ PL_yaourtentremet_freq_M }{  numeric }
  \item{ PL_yaourt_tot_ON }{  character }
  \item{ PL_yaourt_tot_freq_M }{  numeric }
  \item{ PL_yaourt_arom_ON }{  character }
  \item{ PL_yaourt_arom_freq_M }{  numeric }
  \item{ PL_entremet_ON }{  character }
  \item{ PL_entremet_freq_M }{  numeric }
  \item{ PL_yaourtentremet_light_ON }{  character }
  \item{ PL_yaourtentremet_light_freq_M }{  numeric }
  \item{ PL_glace_ON }{  character }
  \item{ PL_glace_freq_M }{  numeric }
  \item{ PL_glace_ete_ON }{  character }
  \item{ PL_glace_ete_freq_M }{  numeric }
  \item{ PL_glace_autresaison_ON }{  character }
  \item{ PL_glace_autresaison_freq_M }{  numeric }
  \item{ FR_total_ON }{  character }
  \item{ FR_total_freq_M }{  numeric }
  \item{ FR_agrume_ON }{  character }
  \item{ FR_agrume_freq_M }{  numeric }
  \item{ FR_agrume_automhiv_ON }{  character }
  \item{ FR_agrume_automhiv_freq_M }{  numeric }
  \item{ FR_agrume_printété_ON }{  character }
  \item{ FR_agrume_printété_freq_M }{  numeric }
  \item{ FR_pomme_ON }{  character }
  \item{ FR_pomme_freq_M }{  numeric }
  \item{ FR_pomme_autresaison_ON }{  character }
  \item{ FR_pomme_autresaison_freq_M }{  numeric }
  \item{ FR_pomme_ete_ON }{  character }
  \item{ FR_pomme_ete_freq_M }{  numeric }
  \item{ FR_raisin_ON }{  character }
  \item{ FR_raisin_freq_M }{  numeric }
  \item{ FR_raisin_automne_ON }{  character }
  \item{ FR_raisin_automne_freq_M }{  numeric }
  \item{ FR_raisin_autresaison_ON }{  character }
  \item{ FR_raisin_autresaison_freq_M }{  numeric }
  \item{ FR_melon_ON }{  character }
  \item{ FR_melon_freq_M }{  numeric }
  \item{ FR_melon_ete_ON }{  character }
  \item{ FR_melon_ete_freq_M }{  numeric }
  \item{ FR_melon_autresaison_ON }{  character }
  \item{ FR_melon_autresaison_freq_M }{  numeric }
  \item{ FR_peche_ON }{  character }
  \item{ FR_peche_freq_M }{  numeric }
  \item{ FR_peche_ete_ON }{  character }
  \item{ FR_peche_ete_freq_M }{  numeric }
  \item{ FR_peche_autresaison_ON }{  character }
  \item{ FR_peche_autresaison_freq_M }{  numeric }
  \item{ FR_fraise_ON }{  character }
  \item{ FR_fraise_freq_M }{  numeric }
  \item{ FR_fraise_printété_ON }{  character }
  \item{ FR_fraise_printété_freq_M }{  numeric }
  \item{ FR_fraise_automhiv_ON }{  character }
  \item{ FR_fraise_automhiv_freq_M }{  numeric }
  \item{ FR_fruitexotique_ON }{  character }
  \item{ FR_fruitexotique_freq_M }{  numeric }
  \item{ BIS_biscuit_ON }{  character }
  \item{ BIS_biscuit_freq_M }{  numeric }
  \item{ BIS_viennoiserie_ON }{  character }
  \item{ BIS_viennoiserie_freq_M }{  numeric }
  \item{ BIS_chocolat_ON }{  character }
  \item{ BIS_chocolat_freq_M }{  numeric }
  \item{ BIS_bonbon_tot_ON }{  character }
  \item{ BIS_bonbon_tot_freq_M }{  numeric }
  \item{ BIS_bonbon_sanssucre_ON }{  character }
  \item{ BIS_bonbon_sanssucre_freq_M }{  numeric }
  \item{ BIS_cacahuete_ON }{  character }
  \item{ BIS_cacahuete_freq_M }{  numeric }
  \item{ BIS_chips_ON }{  character }
  \item{ BIS_chips_freq_M }{  numeric }
  \item{ BNA_jus_tot_ON }{  character }
  \item{ BNA_jus_tot_freq_M }{  numeric }
  \item{ BNA_jus_raisin_ON }{  character }
  \item{ BNA_jus_raisin_freq_M }{  numeric }
  \item{ BNA_sodaredbull_ON }{  character }
  \item{ BNA_sodaredbull_freq_M }{  numeric }
  \item{ BNA_soda_ON }{  character }
  \item{ BNA_soda_freq_M }{  numeric }
  \item{ BNA_redbull_ON }{  character }
  \item{ BNA_redbull_freq_M }{  numeric }
  \item{ BNA_sodaredbull_light_ON }{  character }
  \item{ BNA_sodaredbull_light_freq_M }{  numeric }
  \item{ BA_vin_ON }{  character }
  \item{ BA_vin_freq_M }{  numeric }
  \item{ BA_cidre_ON }{  character }
  \item{ BA_cidre_freq_M }{  numeric }
  \item{ BA_biere_ON }{  character }
  \item{ BA_biere_freq_M }{  numeric }
  \item{ BA_aperitif_ON }{  character }
  \item{ BA_aperitif_freq_M }{  numeric }
  \item{ PC_cereales_bb_ON }{  character }
  \item{ PC_cereales_bb_freq_M }{  numeric }
  \item{ RP_riz_tot_bb_ON }{  character }
  \item{ RP_riz_tot_bb_freq_M }{  numeric }
  \item{ RP_riz_tot_maison_ON }{  character }
  \item{ RP_riz_tot_maison_freq_M }{  numeric }
  \item{ RP_riz_complet_bb_ON }{  character }
  \item{ RP_riz_complet_bb_freq_M }{  numeric }
  \item{ RP_riz_complet_maison_ON }{  character }
  \item{ RP_riz_complet_maison_freq_M }{  numeric }
  \item{ RP_pdt_tot_bb_ON }{  character }
  \item{ RP_pdt_tot_bb_freq_M }{  numeric }
  \item{ RP_pdt_tot_maison_ON }{  character }
  \item{ RP_pdt_tot_maison_freq_M }{  numeric }
  \item{ RP_legumeracine_bb_ON }{  character }
  \item{ RP_legumeracine_bb_freq_M }{  numeric }
  \item{ RP_legumeracine_maison_ON }{  character }
  \item{ RP_legumeracine_maison_freq_M }{  numeric }
  \item{ RP_tubercule_bb_ON }{  character }
  \item{ RP_tubercule_bb_freq_M }{  numeric }
  \item{ RP_tubercule_maison_ON }{  character }
  \item{ RP_tubercule_maison_freq_M }{  numeric }
  \item{ RP_prepariz_bb_ON }{  character }
  \item{ RP_prepariz_bb_freq_M }{  numeric }
  \item{ LEG_tot_bb_ON }{  character }
  \item{ LEG_tot_bb_freq_M }{  numeric }
  \item{ LEG_tot_maison_ON }{  character }
  \item{ LEG_tot_maison_freq_M }{  numeric }
  \item{ LEG_salade_bb_ON }{  character }
  \item{ LEG_salade_bb_freq_M }{  numeric }
  \item{ LEG_salade_maison_ON }{  character }
  \item{ LEG_salade_maison_freq_M }{  numeric }
  \item{ LEG_bettera_tot_bb_ON }{  character }
  \item{ LEG_bettera_tot_bb_freq_M }{  numeric }
  \item{ LEG_bettera_tot_maison_ON }{  character }
  \item{ LEG_bettera_tot_maison_freq_M }{  numeric }
  \item{ LEG_bettera_carot_bb_ON }{  character }
  \item{ LEG_bettera_carot_bb_freq_M }{  numeric }
  \item{ LEG_bettera_carot_maison_ON }{  character }
  \item{ LEG_bettera_carot_maison_freq_M }{  numeric }
  \item{ LEG_haricot_ON }{  character }
  \item{ LEG_haricot_freq_M }{  numeric }
  \item{ LEG_haricot_bb_ON }{  character }
  \item{ LEG_haricot_bb_freq_M }{  numeric }
  \item{ LEG_haricot_maison_ON }{  character }
  \item{ LEG_haricot_maison_freq_M }{  numeric }
  \item{ LEG_tomate_tot_bb_ON }{  character }
  \item{ LEG_tomate_tot_bb_freq_M }{  numeric }
  \item{ LEG_tomate_tot_maison_ON }{  character }
  \item{ LEG_tomate_tot_maison_freq_M }{  numeric }
  \item{ LEG_tomate_seul_bb_ON }{  character }
  \item{ LEG_tomate_seul_bb_freq_M }{  numeric }
  \item{ LEG_tomate_seul_maison_ON }{  character }
  \item{ LEG_tomate_seul_maison_freq_M }{  numeric }
  \item{ LEG_concombre_bb_ON }{  character }
  \item{ LEG_concombre_bb_freq_M }{  numeric }
  \item{ LEG_concombre_maison_ON }{  character }
  \item{ LEG_concombre_maison_freq_M }{  numeric }
  \item{ LEG_chou_bb_ON }{  character }
  \item{ LEG_chou_bb_freq_M }{  numeric }
  \item{ LEG_chou_maison_ON }{  character }
  \item{ LEG_chou_maison_freq_M }{  numeric }
  \item{ LEG_artichaut_bb_ON }{  character }
  \item{ LEG_artichaut_bb_freq_M }{  numeric }
  \item{ LEG_artichaut_maison_ON }{  character }
  \item{ LEG_artichaut_maison_freq_M }{  numeric }
  \item{ LEG_champignon_bb_ON }{  character }
  \item{ LEG_champignon_bb_freq_M }{  numeric }
  \item{ LEG_champignon_maison_ON }{  character }
  \item{ LEG_champignon_maison_freq_M }{  numeric }
  \item{ LEG_legsecs_bb_ON }{  character }
  \item{ LEG_legsecs_bb_freq_M }{  numeric }
  \item{ LEG_legsecs_maison_ON }{  character }
  \item{ LEG_legsecs_maison_freq_M }{  numeric }
  \item{ LEG_soupe_bb_ON }{  character }
  \item{ LEG_soupe_bb_freq_M }{  numeric }
  \item{ LEG_soupe_maison_ON }{  character }
  \item{ LEG_soupe_maison_freq_M }{  numeric }
  \item{ VC_tot_bb_ON }{  character }
  \item{ VC_tot_bb_freq_M }{  numeric }
  \item{ VC_tot_maison_ON }{  character }
  \item{ VC_tot_maison_freq_M }{  numeric }
  \item{ VC_boeuf_bb_ON }{  character }
  \item{ VC_boeuf_bb_freq_M }{  numeric }
  \item{ VC_boeuf_maison_ON }{  character }
  \item{ VC_boeuf_maison_freq_M }{  numeric }
  \item{ VC_agneau_bb_ON }{  character }
  \item{ VC_agneau_bb_freq_M }{  numeric }
  \item{ VC_agneau_maison_ON }{  character }
  \item{ VC_agneau_maison_freq_M }{  numeric }
  \item{ VC_poulet_bb_ON }{  character }
  \item{ VC_poulet_bb_freq_M }{  numeric }
  \item{ VC_poulet_maison_ON }{  character }
  \item{ VC_poulet_maison_freq_M }{  numeric }
  \item{ VC_porc_bb_ON }{  character }
  \item{ VC_porc_bb_freq_M }{  numeric }
  \item{ VC_porc_maison_ON }{  character }
  \item{ VC_porc_maison_freq_M }{  numeric }
  \item{ PPM_tot_bb_ON }{  character }
  \item{ PPM_tot_bb_freq_M }{  numeric }
  \item{ PPM_tot_maison_ON }{  character }
  \item{ PPM_tot_maison_freq_M }{  numeric }
  \item{ PPM_poissgras_tot_bb_ON }{  character }
  \item{ PPM_poissgras_tot_bb_freq_M }{  numeric }
  \item{ PPM_poissgras_tot_maison_ON }{  character }
  \item{ PPM_poissgras_tot_maison_freq_M }{  numeric }
  \item{ PPM_poissonautre_bb_ON }{  character }
  \item{ PPM_poissonautre_bb_freq_M }{  numeric }
  \item{ PPM_poissonautre_maison_ON }{  character }
  \item{ PPM_poissonautre_maison_freq_M }{  numeric }
  \item{ SOJ_prepasoja_bb_ON }{  character }
  \item{ SOJ_prepasoja_bb_freq_M }{  numeric }
  \item{ PL_tot_ON }{  character }
  \item{ PL_tot_freq_M }{  numeric }
  \item{ PL_tot_bb_ON }{  character }
  \item{ PL_tot_bb_freq_M }{  numeric }
  \item{ PL_tot_maison_ON }{  character }
  \item{ PL_tot_maison_freq_M }{  numeric }
  \item{ PL_laitvache_bb_ON }{  character }
  \item{ PL_laitvache_bb_freq_M }{  numeric }
  \item{ PL_laitautre_bb_ON }{  character }
  \item{ PL_laitautre_bb_freq_M }{  logical }
  \item{ PL_laitcereale_bb_ON }{  character }
  \item{ PL_laitcereale_bb_freq_M }{  numeric }
  \item{ PL_yaourt_entremet_ON }{  character }
  \item{ PL_yaourt_entremet_freq_M }{  numeric }
  \item{ PL_yaourt_tot_bb_ON }{  character }
  \item{ PL_yaourt_tot_bb_freq_M }{  numeric }
  \item{ PL_yaourt_tot_maison_ON }{  character }
  \item{ PL_yaourt_tot_maison_freq_M }{  numeric }
  \item{ PL_yaourt_arom_bb_ON }{  character }
  \item{ PL_yaourt_arom_bb_freq_M }{  numeric }
  \item{ PL_yaourt_arom_maison_ON }{  character }
  \item{ PL_yaourt_arom_maison_freq_M }{  numeric }
  \item{ PL_entremet_bb_ON }{  character }
  \item{ PL_entremet_bb_freq_M }{  numeric }
  \item{ PL_entremet_maison_ON }{  character }
  \item{ PL_entremet_maison_freq_M }{  numeric }
  \item{ FR_total_bb_ON }{  character }
  \item{ FR_total_bb_freq_M }{  numeric }
  \item{ FR_total_maison_ON }{  character }
  \item{ FR_total_maison_freq_M }{  numeric }
  \item{ FR_banane_ON }{  character }
  \item{ FR_banane_freq_M }{  numeric }
  \item{ FR_banane_bb_ON }{  character }
  \item{ FR_banane_bb_freq_M }{  numeric }
  \item{ FR_banane_maison_ON }{  character }
  \item{ FR_banane_maison_freq_M }{  numeric }
  \item{ FR_pomme_bb_ON }{  character }
  \item{ FR_pomme_bb_freq_M }{  numeric }
  \item{ FR_pomme_maison_ON }{  character }
  \item{ FR_pomme_maison_freq_M }{  numeric }
  \item{ FR_raisin_bb_ON }{  character }
  \item{ FR_raisin_bb_freq_M }{  numeric }
  \item{ FR_raisin_maison_ON }{  character }
  \item{ FR_raisin_maison_freq_M }{  numeric }
  \item{ FR_peche_bb_ON }{  character }
  \item{ FR_peche_bb_freq_M }{  numeric }
  \item{ FR_peche_maison_ON }{  character }
  \item{ FR_peche_maison_freq_M }{  numeric }
  \item{ FR_fruitrouge_ON }{  character }
  \item{ FR_fruitrouge_freq_M }{  numeric }
  \item{ FR_fruitrouge_bb_ON }{  character }
  \item{ FR_fruitrouge_bb_freq_M }{  numeric }
  \item{ FR_fruitrouge_maison_ON }{  character }
  \item{ FR_fruitrouge_maison_freq_M }{  numeric }
  \item{ FR_fruitexotique_bb_ON }{  character }
  \item{ FR_fruitexotique_bb_freq_M }{  numeric }
  \item{ FR_fruitexotique_maison_ON }{  character }
  \item{ FR_fruitexotique_maison_freq_M }{  numeric }
  \item{ BIS_biscuit_bb_ON }{  character }
  \item{ BIS_biscuit_bb_freq_M }{  numeric }
  \item{ BIS_biscuit_maison_ON }{  character }
  \item{ BIS_biscuit_maison_freq_M }{  numeric }
  \item{ BNA_jus_tot_bb_ON }{  character }
  \item{ BNA_jus_tot_bb_freq_M }{  numeric }
  \item{ BNA_jus_tot_maison_ON }{  character }
  \item{ BNA_jus_tot_maison_freq_M }{  numeric }
  \item{ BNA_jus_raisin_bb_ON }{  character }
  \item{ BNA_jus_raisin_bb_freq_M }{  numeric }
  \item{ BNA_jus_raisin_maison_ON }{  character }
  \item{ BNA_jus_raisin_maison_freq_M }{  numeric }
}
}
\source{
Inca3
}
\usage{
fpq_decode
}
\description{
Data inca3.
}
\keyword{datasets}
