% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitudes_indiv.R
\docType{data}
\name{habitudes_indiv}
\alias{habitudes_indiv}
\title{habitudes_indiv}
\format{A data frame with 4372 rows and 389 variables:
\describe{
  \item{ POPULATION }{  character }
  \item{ NOIND }{  numeric }
  \item{ periode_reference }{  numeric }
  \item{ repasenfantmidi_vous }{  numeric }
  \item{ repasenfantmidi_am }{  numeric }
  \item{ repasenfantmidi_creche }{  numeric }
  \item{ repasenfantmidi_parent }{  numeric }
  \item{ repasenfantmidi_autre }{  numeric }
  \item{ repasenfantmidi_autre_libelle }{  logical }
  \item{ repasenfant_freq }{  numeric }
  \item{ conso_bebe_sein }{  numeric }
  \item{ lait_plussouvent }{  numeric }
  \item{ lait_plusvt_li_autr_anim_libelle }{  logical }
  \item{ lait_plusvt_pi_autre_veg_libelle }{  logical }
  \item{ lait_plusvt_boisson_veg_libelle }{  character }
  \item{ lait_plusvt_autre_anim_libelle }{  logical }
  \item{ lait_plusvt_autre_libelle }{  logical }
  \item{ lait_occ_laitmaternel }{  numeric }
  \item{ lait_occ_lait_1e_age }{  numeric }
  \item{ lait_occ_lait_2e_age }{  numeric }
  \item{ lait_occ_lait_croissance }{  numeric }
  \item{ lait_occ_li_autre_anim }{  numeric }
  \item{ lait_occ_li_autre_anim_libelle }{  logical }
  \item{ lait_occ_li1e_age_protsoja }{  numeric }
  \item{ lait_occ_li2e_age_protsoja }{  numeric }
  \item{ lait_occ_soja_croissance }{  numeric }
  \item{ lait_occ_li1e_age_protriz }{  numeric }
  \item{ lait_occ_li2e_age_protriz }{  numeric }
  \item{ lait_occ_riz_croissance }{  numeric }
  \item{ lait_occ_pi_autre_veg }{  numeric }
  \item{ lait_occ_pi_autre_veg_libelle }{  logical }
  \item{ lait_occ_boisson_lait_bebe }{  numeric }
  \item{ lait_occ_lait_regime_infantile }{  numeric }
  \item{ lait_occ_lait_vache }{  numeric }
  \item{ lait_occ_boisson_soja }{  numeric }
  \item{ lait_occ_boisson_veg }{  numeric }
  \item{ lait_occ_boisson_veg_libelle }{  logical }
  \item{ lait_occ_autre_anim }{  numeric }
  \item{ lait_occ_autre_anim_libelle }{  logical }
  \item{ lait_occ_autre }{  numeric }
  \item{ lait_occ_autre_libelle }{  logical }
  \item{ lait_occ_aucun }{  numeric }
  \item{ lait_occ_nsp }{  numeric }
  \item{ etablissement_scolaire }{  numeric }
  \item{ lieu_repas_midi }{  numeric }
  \item{ lieu_repas_midi_autre_libelle }{  character }
  \item{ cantine_freq }{  numeric }
  \item{ cantine_jms_persfoyer }{  numeric }
  \item{ cantine_jms_veutpas }{  numeric }
  \item{ cantine_jms_probsante }{  numeric }
  \item{ cantine_jms_repaspasequilibre }{  numeric }
  \item{ cantine_jms_repaspasqualite }{  numeric }
  \item{ cantine_jms_repaschers }{  numeric }
  \item{ cantine_jms_ecoleproche }{  numeric }
  \item{ cantine_jms_paspropose }{  numeric }
  \item{ cantine_jms_pasaime }{  numeric }
  \item{ cantine_jms_autre_raison }{  numeric }
  \item{ cantine_jms_autre_raison_libelle }{  character }
  \item{ cantine_jms_nsp }{  numeric }
  \item{ collation_freq }{  numeric }
  \item{ collation_aliment1 }{  character }
  \item{ collation_aliment2 }{  character }
  \item{ collation_aliment3 }{  character }
  \item{ collation_aliment4 }{  character }
  \item{ collation_matin_freq }{  numeric }
  \item{ collation_matin_ecole_freq }{  numeric }
  \item{ collation_matin_pasecole_freq }{  numeric }
  \item{ collation_matin_aliment1 }{  character }
  \item{ collation_matin_aliment2 }{  character }
  \item{ collation_matin_aliment3 }{  character }
  \item{ collation_matin_aliment4 }{  character }
  \item{ distributeur_freq }{  numeric }
  \item{ distributeur_biscuits_sales }{  numeric }
  \item{ distributeur_biscuits_sucres }{  numeric }
  \item{ distributeur_barres_choco }{  numeric }
  \item{ distributeur_confiseries }{  numeric }
  \item{ distributeur_fruits_frais }{  numeric }
  \item{ distributeur_jus_de_fruits }{  numeric }
  \item{ distributeur_boissons_energ }{  numeric }
  \item{ distributeur_sodas }{  numeric }
  \item{ distributeur_eaux }{  numeric }
  \item{ distributeur_autre }{  numeric }
  \item{ distributeur_autre_libelle }{  character }
  \item{ distributeur_nsp }{  numeric }
  \item{ restaurationrapide_freq }{  numeric }
  \item{ aime_legumes }{  numeric }
  \item{ aime_viande }{  numeric }
  \item{ aime_fruits }{  numeric }
  \item{ aime_glaces }{  numeric }
  \item{ aime_fromage }{  numeric }
  \item{ aime_poisson }{  numeric }
  \item{ aime_soda }{  numeric }
  \item{ aime_lait }{  numeric }
  \item{ aime_yaourts }{  numeric }
  \item{ aime_jus_de_fruits }{  numeric }
  \item{ table_beurre }{  numeric }
  \item{ table_creme_fraiche }{  numeric }
  \item{ table_huile_olive }{  numeric }
  \item{ table_vinaigrette }{  numeric }
  \item{ table_mayonnaise }{  numeric }
  \item{ table_ketchup }{  numeric }
  \item{ table_sauce_soja }{  numeric }
  \item{ table_sel }{  numeric }
  \item{ etiquette_freq }{  numeric }
  \item{ etiquette_ingredients }{  numeric }
  \item{ etiquette_contenu_nutri }{  numeric }
  \item{ etiquette_portions }{  numeric }
  \item{ etiquette_message_nutri }{  numeric }
  \item{ etiquette_effets_benefiques }{  numeric }
  \item{ source_medecins }{  numeric }
  \item{ source_personnels_pharma }{  numeric }
  \item{ source_publicite }{  numeric }
  \item{ source_emissions_info }{  numeric }
  \item{ source_journaux }{  numeric }
  \item{ source_livres }{  numeric }
  \item{ source_professeurs }{  numeric }
  \item{ source_famille }{  numeric }
  \item{ source_amis }{  numeric }
  \item{ source_internet }{  numeric }
  \item{ source_emballages }{  numeric }
  \item{ source_autre }{  numeric }
  \item{ source_autre_libelle }{  character }
  \item{ conso_lardons_crus }{  numeric }
  \item{ conso_saucisses_crues }{  numeric }
  \item{ conso_viande_boeuf_crue }{  numeric }
  \item{ conso_viande_cheval_crue }{  numeric }
  \item{ conso_viande_porc_crue }{  numeric }
  \item{ conso_viande_volaille_crue }{  numeric }
  \item{ conso_poisson_cru }{  numeric }
  \item{ conso_mollusques_crus }{  numeric }
  \item{ conso_oeufs_crus }{  numeric }
  \item{ conso_croute_fleurie }{  numeric }
  \item{ conso_croute_lavee }{  numeric }
  \item{ conso_pate_pressee_non_cuite }{  numeric }
  \item{ conso_pate_pressee_cuite }{  numeric }
  \item{ conso_pate_persillee }{  numeric }
  \item{ pomme_crue_lave }{  numeric }
  \item{ pomme_crue_essuie }{  numeric }
  \item{ pomme_crue_epluche }{  numeric }
  \item{ pomme_crue_rien }{  numeric }
  \item{ pomme_crue_mange_pas }{  numeric }
  \item{ peche_crue_lave }{  numeric }
  \item{ peche_crue_essuie }{  numeric }
  \item{ peche_crue_epluche }{  numeric }
  \item{ peche_crue_rien }{  numeric }
  \item{ peche_crue_mange_pas }{  numeric }
  \item{ raisin_cru_lave }{  numeric }
  \item{ raisin_cru_essuie }{  numeric }
  \item{ raisin_cru_epluche }{  numeric }
  \item{ raisin_cru_rien }{  numeric }
  \item{ raisin_cru_mange_pas }{  numeric }
  \item{ fraise_crue_lave }{  numeric }
  \item{ fraise_crue_essuie }{  numeric }
  \item{ fraise_crue_epluche }{  numeric }
  \item{ fraise_crue_rien }{  numeric }
  \item{ fraise_crue_mange_pas }{  numeric }
  \item{ tomate_crue_lave }{  numeric }
  \item{ tomate_crue_essuie }{  numeric }
  \item{ tomate_crue_epluche }{  numeric }
  \item{ tomate_crue_rien }{  numeric }
  \item{ tomate_crue_mange_pas }{  numeric }
  \item{ champignon_cru_lave }{  numeric }
  \item{ champignon_cru_essuie }{  numeric }
  \item{ champignon_cru_epluche }{  numeric }
  \item{ champignon_cru_rien }{  numeric }
  \item{ champignon_cru_mange_pas }{  numeric }
  \item{ salade_crue_lave }{  numeric }
  \item{ salade_crue_essuie }{  numeric }
  \item{ salade_crue_epluche }{  numeric }
  \item{ salade_crue_rien }{  numeric }
  \item{ salade_crue_mange_pas }{  numeric }
  \item{ concombre_cru_lave }{  numeric }
  \item{ concombre_cru_essuie }{  numeric }
  \item{ concombre_cru_epluche }{  numeric }
  \item{ concombre_cru_rien }{  numeric }
  \item{ concombre_cru_mange_pas }{  numeric }
  \item{ utilisation_bib }{  numeric }
  \item{ bib_plastique }{  numeric }
  \item{ bib_verre }{  numeric }
  \item{ bib_nsp }{  numeric }
  \item{ sterilisation_bib }{  numeric }
  \item{ sterilisation_bib_libelle }{  character }
  \item{ tetine_caoutchouc }{  numeric }
  \item{ tetine_silicone }{  numeric }
  \item{ tetine_nsp }{  numeric }
  \item{ sterilisation_tetine }{  numeric }
  \item{ sterilisation_tetine_libelle }{  character }
  \item{ bib_lm }{  numeric }
  \item{ bib_li }{  numeric }
  \item{ bib_lc }{  numeric }
  \item{ bib_bl }{  numeric }
  \item{ bib_ppl }{  numeric }
  \item{ bib_eau }{  numeric }
  \item{ bib_jus }{  numeric }
  \item{ bib_autre }{  numeric }
  \item{ bib_autre_libelle }{  logical }
  \item{ stockage_bib_lm }{  numeric }
  \item{ delai_conso_bib_lm }{  numeric }
  \item{ prepa_bib_li }{  numeric }
  \item{ conserv_eau }{  numeric }
  \item{ conserv_eau_libelle }{  logical }
  \item{ reste_eau }{  numeric }
  \item{ conso_apres_prepa_bib_li }{  numeric }
  \item{ tps_conserv_bib_li }{  numeric }
  \item{ lieu_conserv_bib_li }{  numeric }
  \item{ lieu_conserv_bib_li_libelle }{  logical }
  \item{ tps_conso_bib_li }{  numeric }
  \item{ chauffe_lm }{  numeric }
  \item{ chauffe_pi }{  numeric }
  \item{ chauffe_lc }{  numeric }
  \item{ chauffe_bl }{  numeric }
  \item{ bib_micro_ondes }{  numeric }
  \item{ bib_chauffe_bib }{  numeric }
  \item{ bib_bainmarie }{  numeric }
  \item{ diversification_alim_bebe }{  numeric }
  \item{ prepbb }{  numeric }
  \item{ prepbb_chauff }{  numeric }
  \item{ prepbb_chauff_coupplastique }{  numeric }
  \item{ prepbb_chauff_potverre }{  numeric }
  \item{ prepbb_chauff_recipopaque }{  numeric }
  \item{ prepbb_chauff_reciptranslucid }{  numeric }
  \item{ prepbb_chauff_recipceramique }{  numeric }
  \item{ prepbb_chauff_recipverre }{  numeric }
  \item{ prepbb_chauff_autre }{  numeric }
  \item{ prepbb_chauff_autre_libelle }{  logical }
  \item{ conso_farine_ssgluten }{  numeric }
  \item{ conso_farine_ssgluten_age }{  numeric }
  \item{ conso_farine_ssgluten_age_nsp }{  numeric }
  \item{ conso_farine_acgluten }{  numeric }
  \item{ conso_farine_acgluten_age }{  numeric }
  \item{ conso_farine_acgluten_age_nsp }{  numeric }
  \item{ conso_cereales }{  numeric }
  \item{ conso_cereales_age }{  numeric }
  \item{ conso_cereales_age_nsp }{  numeric }
  \item{ conso_legumes_horspdt }{  numeric }
  \item{ conso_legumes_horspdt_age }{  numeric }
  \item{ conso_legumes_horspdt_age_nsp }{  numeric }
  \item{ conso_pdt }{  numeric }
  \item{ conso_pdt_age }{  numeric }
  \item{ conso_pdt_age_nsp }{  numeric }
  \item{ conso_riz_pates }{  numeric }
  \item{ conso_riz_pates_age }{  numeric }
  \item{ conso_riz_pates_age_nsp }{  numeric }
  \item{ conso_fruits }{  numeric }
  \item{ conso_fruits_age }{  numeric }
  \item{ conso_fruits_age_nsp }{  numeric }
  \item{ conso_yaourts }{  numeric }
  \item{ conso_yaourts_age }{  numeric }
  \item{ conso_yaourts_age_nsp }{  numeric }
  \item{ conso_autres_desserts }{  numeric }
  \item{ conso_autres_desserts_age }{  numeric }
  \item{ conso_autres_desserts_age_nsp }{  numeric }
  \item{ conso_fromage }{  numeric }
  \item{ conso_fromage_age }{  numeric }
  \item{ conso_fromage_age_nsp }{  numeric }
  \item{ conso_viande }{  numeric }
  \item{ conso_viande_age }{  numeric }
  \item{ conso_viande_age_nsp }{  numeric }
  \item{ conso_poisson }{  numeric }
  \item{ conso_poisson_age }{  numeric }
  \item{ conso_poisson_age_nsp }{  numeric }
  \item{ conso_jambon }{  numeric }
  \item{ conso_jambon_age }{  numeric }
  \item{ conso_jambon_age_nsp }{  numeric }
  \item{ conso_oeufs }{  numeric }
  \item{ conso_oeufs_age }{  numeric }
  \item{ conso_oeufs_age_nsp }{  numeric }
  \item{ conso_gateaux }{  numeric }
  \item{ conso_gateaux_age }{  numeric }
  \item{ conso_gateaux_age_nsp }{  numeric }
  \item{ conso_pain }{  numeric }
  \item{ conso_pain_age }{  numeric }
  \item{ conso_pain_age_nsp }{  numeric }
  \item{ conso_jus }{  numeric }
  \item{ conso_jus_age }{  numeric }
  \item{ conso_jus_age_nsp }{  numeric }
  \item{ conso_autres }{  numeric }
  \item{ conso_autres_age }{  logical }
  \item{ conso_autres_age_nsp }{  numeric }
  \item{ conso_autres_libelle }{  logical }
  \item{ conso_nsp }{  logical }
  \item{ conso_plats_faits_maison }{  numeric }
  \item{ jardin }{  numeric }
  \item{ jardin_potager }{  numeric }
  \item{ jardin_arbres_fruitiers }{  numeric }
  \item{ jardin_poulailler }{  numeric }
  \item{ jardin_clapier }{  numeric }
  \item{ jardin_fleurs }{  numeric }
  \item{ jardin_rien }{  numeric }
  \item{ jardin_entretien }{  numeric }
  \item{ autoproduction_aliments }{  numeric }
  \item{ autoproduction_oeufs_freq }{  numeric }
  \item{ autoproduction_volailles_freq }{  numeric }
  \item{ autoproduction_volailles_libelle }{  character }
  \item{ autoproduction_lapins_freq }{  numeric }
  \item{ autoproduction_viandes_freq }{  numeric }
  \item{ autoproduction_viandes_libelle }{  character }
  \item{ autoproduction_lait_freq }{  numeric }
  \item{ autoproduction_pdt_freq }{  numeric }
  \item{ autoproduction_tomates_freq }{  numeric }
  \item{ autoproduction_tomates_libelle }{  character }
  \item{ autoproduction_carottes_freq }{  numeric }
  \item{ autoproduction_carottes_libelle }{  character }
  \item{ autoproduction_salades_freq }{  numeric }
  \item{ autoproduction_salades_libelle }{  character }
  \item{ autoproduction_abricots_freq }{  numeric }
  \item{ autoproduction_abricots_libelle }{  character }
  \item{ autoproduction_pommes_freq }{  numeric }
  \item{ autoproduction_pommes_libelle }{  character }
  \item{ autoproduction_fraises_freq }{  numeric }
  \item{ autoproduction_fraises_libelle }{  character }
  \item{ autoproduction_raisins_freq }{  numeric }
  \item{ autoproduction_autresfruits_freq }{  numeric }
  \item{ autoproduction_autrfruit_libelle }{  character }
  \item{ autoproduction_autreslegum_freq }{  numeric }
  \item{ autoproduction_autrlegum_libelle }{  character }
  \item{ cueillette_fruits_freq }{  numeric }
  \item{ cueillette_champignons_freq }{  numeric }
  \item{ peche_poissons_eau_douce_freq }{  numeric }
  \item{ peche_poissons_eau_mer_freq }{  numeric }
  \item{ peche_coquillages_freq }{  numeric }
  \item{ chasse_gibiers_freq }{  numeric }
  \item{ autoconsommation }{  numeric }
  \item{ autoconsommation_type }{  numeric }
  \item{ autoproduction }{  numeric }
  \item{ cueillette }{  numeric }
  \item{ autoconso_oeufs }{  numeric }
  \item{ autoconso_oeufs_type }{  numeric }
  \item{ autoconso_viandes }{  numeric }
  \item{ autoconso_viandes_type }{  numeric }
  \item{ autoconso_volailles_tot }{  numeric }
  \item{ autoconso_volailles_tot_type }{  numeric }
  \item{ autoconso_lait }{  numeric }
  \item{ autoconso_lait_type }{  numeric }
  \item{ autoconso_pdt }{  numeric }
  \item{ autoconso_pdt_type }{  numeric }
  \item{ autoconso_legumes_tot }{  numeric }
  \item{ autoconso_legumes_tot_type }{  numeric }
  \item{ autoconso_fruits_tot }{  numeric }
  \item{ autoconso_fruits_tot_type }{  numeric }
  \item{ autoconso_pdts_mer }{  numeric }
  \item{ autoconso_pdts_mer_type }{  numeric }
  \item{ conso_aliments_bio }{  numeric }
  \item{ conso_lait_bio_freq }{  numeric }
  \item{ conso_oeufs_bio_freq }{  numeric }
  \item{ conso_volailles_bio_freq }{  numeric }
  \item{ conso_viandes_bio_freq }{  numeric }
  \item{ conso_poissons_bio_freq }{  numeric }
  \item{ conso_fruits_bio_freq }{  numeric }
  \item{ conso_fruits_bio_libelle }{  character }
  \item{ conso_pdt_bio_freq }{  numeric }
  \item{ conso_legumes_bio_freq }{  numeric }
  \item{ conso_legumes_bio_libelle }{  character }
  \item{ conso_legumes_secs_bio_freq }{  numeric }
  \item{ conso_legumes_secs_bio_libelle }{  character }
  \item{ conso_pain_bio_freq }{  numeric }
  \item{ conso_cereales_bio_freq }{  numeric }
  \item{ conso_cereales_bio_libelle }{  character }
  \item{ conso_autres_alim_bio_freq }{  numeric }
  \item{ conso_autres_alim_bio_libelle }{  character }
  \item{ conso_lait_bebe_bio_freq }{  numeric }
  \item{ conso_yaourt_bebe_bio_freq }{  numeric }
  \item{ conso_jus_bebe_bio_freq }{  numeric }
  \item{ conso_puree_bebe_bio_freq }{  numeric }
  \item{ conso_cereales_bebe_bio_freq }{  numeric }
  \item{ conso_biscuits_bebe_bio_freq }{  numeric }
  \item{ conso_plats_bebe_bio_freq }{  numeric }
  \item{ conso_compotes_bebe_bio_freq }{  numeric }
  \item{ conso_cremes_bebe_bio_freq }{  numeric }
  \item{ consommation_bio }{  numeric }
  \item{ consommation_bio_type }{  numeric }
  \item{ conso_oeufs_bio_tjr_svt }{  numeric }
  \item{ conso_lait_bio_tjr_svt }{  numeric }
  \item{ conso_volailles_bio_tjr_svt }{  numeric }
  \item{ conso_viandes_bio_tjr_svt }{  numeric }
  \item{ conso_poissons_bio_tjr_svt }{  numeric }
  \item{ conso_fruits_bio_tjr_svt }{  numeric }
  \item{ conso_pdt_bio_tjr_svt }{  numeric }
  \item{ conso_legumes_bio_tjr_svt }{  numeric }
  \item{ conso_legumes_secs_bio_tjr_svt }{  numeric }
  \item{ conso_cereales_bio_tjr_svt }{  numeric }
  \item{ conso_pain_bio_tjr_svt }{  numeric }
  \item{ conso_autres_alim_bio_tjr_svt }{  numeric }
  \item{ conso_barbecue }{  numeric }
  \item{ conso_barbecue_elec }{  numeric }
  \item{ conso_barbecue_bois }{  numeric }
  \item{ conso_barbecue_gaz }{  numeric }
  \item{ conso_barbecue_printps_ete_freq }{  numeric }
  \item{ conso_barbecue_automn_hiver_freq }{  numeric }
}}
\source{
Inca3
}
\usage{
habitudes_indiv
}
\description{
Data inca3.
}
\keyword{datasets}
