% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitudes_men.R
\docType{data}
\name{habitudes_men}
\alias{habitudes_men}
\title{habitudes_men}
\format{A data frame with 4372 rows and 167 variables:
\describe{
  \item{ POPULATION }{  character }
  \item{ NOMEN }{  numeric }
  \item{ pomme_cuite_lave }{  numeric }
  \item{ pomme_cuite_essuie }{  numeric }
  \item{ pomme_cuite_epluche }{  numeric }
  \item{ pomme_cuite_rien }{  numeric }
  \item{ pomme_cuite_cuisine_pas }{  numeric }
  \item{ peche_cuite_lave }{  numeric }
  \item{ peche_cuite_essuie }{  numeric }
  \item{ peche_cuite_epluche }{  numeric }
  \item{ peche_cuite_rien }{  numeric }
  \item{ peche_cuite_cuisine_pas }{  numeric }
  \item{ raisin_cuit_lave }{  numeric }
  \item{ raisin_cuit_essuie }{  numeric }
  \item{ raisin_cuit_epluche }{  numeric }
  \item{ raisin_cuit_rien }{  numeric }
  \item{ raisin_cuit_cuisine_pas }{  numeric }
  \item{ fraise_cuite_lave }{  numeric }
  \item{ fraise_cuite_essuie }{  numeric }
  \item{ fraise_cuite_epluche }{  numeric }
  \item{ fraise_cuite_rien }{  numeric }
  \item{ fraise_cuite_cuisine_pas }{  numeric }
  \item{ tomate_cuite_lave }{  numeric }
  \item{ tomate_cuite_essuie }{  numeric }
  \item{ tomate_cuite_epluche }{  numeric }
  \item{ tomate_cuite_rien }{  numeric }
  \item{ tomate_cuite_cuisine_pas }{  numeric }
  \item{ champignon_cuit_lave }{  numeric }
  \item{ champignon_cuit_essuie }{  numeric }
  \item{ champignon_cuit_epluche }{  numeric }
  \item{ champignon_cuit_rien }{  numeric }
  \item{ champignon_cuit_cuisine_pas }{  numeric }
  \item{ chou_cuit_lave }{  numeric }
  \item{ chou_cuit_essuie }{  numeric }
  \item{ chou_cuit_epluche }{  numeric }
  \item{ chou_cuit_rien }{  numeric }
  \item{ chou_cuit_cuisine_pas }{  numeric }
  \item{ concombre_cuit_lave }{  numeric }
  \item{ concombre_cuit_essuie }{  numeric }
  \item{ concombre_cuit_epluche }{  numeric }
  \item{ concombre_cuit_rien }{  numeric }
  \item{ concombre_cuit_cuisine_pas }{  numeric }
  \item{ recip_cuisson_inox }{  numeric }
  \item{ recip_cuisson_alu }{  numeric }
  \item{ recip_cuisson_metal }{  numeric }
  \item{ recip_cuisson_ceramique }{  numeric }
  \item{ recip_cuisson_fonte }{  numeric }
  \item{ recip_cuisson_antiadhesif }{  numeric }
  \item{ recip_cuisson_autre }{  numeric }
  \item{ recip_cuisson_autre_libelle }{  character }
  \item{ recip_cuisson_nsp }{  numeric }
  \item{ four_ondes }{  numeric }
  \item{ four_ondes_vaisselle_maison }{  numeric }
  \item{ four_ondes_vaiss_maison_libelle }{  logical }
  \item{ four_ondes_vaisselle }{  numeric }
  \item{ four_ondes_vaiss_libelle }{  logical }
  \item{ choix_aliment_origine }{  numeric }
  \item{ choix_aliment_prix }{  numeric }
  \item{ choix_aliment_offres_promo }{  numeric }
  \item{ choix_aliment_habitude }{  numeric }
  \item{ choix_aliment_marque }{  numeric }
  \item{ choix_aliment_facilite_prepa }{  numeric }
  \item{ choix_aliment_apparence }{  numeric }
  \item{ choix_aliment_mode_prod }{  numeric }
  \item{ choix_aliment_gout }{  numeric }
  \item{ choix_aliment_prod_nouveau }{  numeric }
  \item{ choix_aliment_pub }{  numeric }
  \item{ choix_aliment_signe_quali }{  numeric }
  \item{ choix_aliment_compo_nutri }{  numeric }
  \item{ choix_aliment_ingredients }{  numeric }
  \item{ choix_aliment_raison_sante }{  numeric }
  \item{ choix_aliment_autres_infos }{  numeric }
  \item{ choix_prod_identiques }{  numeric }
  \item{ lieu_achat_fruits_frais }{  numeric }
  \item{ lieu_achat_legumes_frais }{  numeric }
  \item{ lieu_achat_pain }{  numeric }
  \item{ lieu_achat_patisseries }{  numeric }
  \item{ lieu_achat_viandes }{  numeric }
  \item{ lieu_achat_poissons_frais }{  numeric }
  \item{ suivi_instruction_conserv }{  numeric }
  \item{ refroidit_viandes }{  numeric }
  \item{ refroidit_quiches }{  numeric }
  \item{ refroidit_legumes }{  numeric }
  \item{ refroidit_soupes }{  numeric }
  \item{ refroidit_feculents }{  numeric }
  \item{ refroidit_cremesdessert }{  numeric }
  \item{ refroidit_patisseries }{  numeric }
  \item{ fin_conso_platmaison }{  numeric }
  \item{ conso_dlc_jambon }{  numeric }
  \item{ conso_dlc_saumon }{  numeric }
  \item{ conso_dlc_platscuisines }{  numeric }
  \item{ conso_dlc_viandepreemballee }{  numeric }
  \item{ conso_dlc_beurre }{  numeric }
  \item{ conso_apresachat_jamboncoupe }{  numeric }
  \item{ conso_apresachat_autrecharcut }{  numeric }
  \item{ conso_apresachat_platscuisines }{  numeric }
  \item{ conso_apresachat_entreesfroides }{  numeric }
  \item{ conso_apresachat_fromagecoupe }{  numeric }
  \item{ conso_apresachat_viande }{  numeric }
  \item{ temp_frigo }{  numeric }
  \item{ partie_froide_frigo }{  numeric }
  \item{ eaupuits_utilisation }{  numeric }
  \item{ eaupuits_boisson }{  numeric }
  \item{ eaupuits_prepa_repas }{  numeric }
  \item{ eaupuits_arr_jardin }{  numeric }
  \item{ eaupuits_arr_potager }{  numeric }
  \item{ eaupuits_piscine }{  numeric }
  \item{ eaupuits_toilette }{  numeric }
  \item{ eaupuits_wc }{  numeric }
  \item{ eaupuits_lave_linge }{  numeric }
  \item{ eaupuits_boisson_animaux }{  numeric }
  \item{ eaupuits_nettoyage }{  numeric }
  \item{ eaupuits_rien }{  numeric }
  \item{ eaupuits_autre }{  numeric }
  \item{ eaupuits_autre_libelle }{  character }
  \item{ eaupuits_nsp }{  numeric }
  \item{ eaupuits_conso_enfant }{  logical }
  \item{ eaupuits_traitement }{  numeric }
  \item{ eaurobinet }{  numeric }
  \item{ eaurobinet_conso }{  numeric }
  \item{ eaurobinet_traitement }{  numeric }
  \item{ eaurobinet_trait_circeau }{  numeric }
  \item{ eaurobinet_trait_toutcirc }{  numeric }
  \item{ eaurobinet_trait_ptusage }{  numeric }
  \item{ eaurobinet_trait_ptconso }{  numeric }
  \item{ eaurobinet_trait_nsp }{  numeric }
  \item{ trait_carafe }{  numeric }
  \item{ trait_carafe_marque_libelle }{  character }
  \item{ trait_carafe_gout }{  numeric }
  \item{ trait_carafe_contaminant }{  numeric }
  \item{ trait_carafe_bacteries }{  numeric }
  \item{ trait_carafe_entartrage }{  numeric }
  \item{ trait_adoucisseur }{  numeric }
  \item{ trait_adoucisseur_gout }{  numeric }
  \item{ trait_adoucisseur_contaminant }{  numeric }
  \item{ trait_adoucisseur_bacteries }{  numeric }
  \item{ trait_adoucisseur_entartrage }{  numeric }
  \item{ trait_cartouche }{  numeric }
  \item{ trait_cartouche_gout }{  numeric }
  \item{ trait_cartouche_contaminant }{  numeric }
  \item{ trait_cartouche_bacteries }{  numeric }
  \item{ trait_cartouche_entartrage }{  numeric }
  \item{ trait_ultraviolet }{  numeric }
  \item{ trait_ultraviolet_gout }{  numeric }
  \item{ trait_ultraviolet_contaminant }{  numeric }
  \item{ trait_ultraviolet_bacteries }{  numeric }
  \item{ trait_ultraviolet_entartrage }{  numeric }
  \item{ trait_cart_refrig }{  numeric }
  \item{ trait_cart_refrig_gout }{  numeric }
  \item{ trait_cart_refrig_contaminant }{  numeric }
  \item{ trait_cart_refrig_bacteries }{  numeric }
  \item{ trait_cart_refrig_entartrage }{  numeric }
  \item{ trait_osmoseur }{  numeric }
  \item{ trait_osmoseur_gout }{  numeric }
  \item{ trait_osmoseur_contaminant }{  numeric }
  \item{ trait_osmoseur_bacteries }{  numeric }
  \item{ trait_osmoseur_entartrage }{  numeric }
  \item{ trait_autre }{  numeric }
  \item{ trait_autre_libelle }{  character }
  \item{ trait_autre_gout }{  numeric }
  \item{ trait_autre_contaminant }{  numeric }
  \item{ trait_autre_bacteries }{  numeric }
  \item{ trait_autre_entartrage }{  numeric }
  \item{ trait_nsp }{  numeric }
  \item{ eau_boisson }{  numeric }
  \item{ bouilloire_elec }{  numeric }
  \item{ bouilloire_elec_type }{  numeric }
}}
\source{
Inca3
}
\usage{
habitudes_men
}
\description{
Data inca3.
}
\keyword{datasets}
