% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liaison_format.R
\docType{data}
\name{liaison_format}
\alias{liaison_format}
\title{liaison_format}
\format{
A data frame with 1577 rows and 4 variables:
\describe{
  \item{ VARIABLE }{  character }
  \item{ TYPE }{  character }
  \item{ FORMAT }{  character }
  \item{ LIBELLE }{  character }
}
}
\source{
Inca3
}
\usage{
liaison_format
}
\description{
Data inca3.
}
\keyword{datasets}
