% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occasions.R
\docType{data}
\name{occasions}
\alias{occasions}
\title{occasions}
\format{
A data frame with 131097 rows and 13 variables:
\describe{
  \item{ POPULATION }{  character }
  \item{ NOIND }{  numeric }
  \item{ R24_num }{  numeric }
  \item{ R24_jour }{  numeric }
  \item{ R24_nombre }{  numeric }
  \item{ sem_we }{  numeric }
  \item{ sem_we_ferie }{  numeric }
  \item{ occ_ouinon }{  numeric }
  \item{ occ_hdeb }{  hms,difftime }
  \item{ occ_type }{  numeric }
  \item{ occ_lieu }{  numeric }
  \item{ occ_lieu_autre_libelle }{  character }
  \item{ occ_lieu_dom_hors_dom }{  numeric }
}
}
\source{
Inca3
}
\usage{
occasions
}
\description{
Data inca3.
}
\keyword{datasets}
