% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occasions_decode.R
\docType{data}
\name{occasions_decode}
\alias{occasions_decode}
\title{occasions_decode}
\format{A data frame with 131097 rows and 13 variables:
\describe{
  \item{ POPULATION }{  character }
  \item{ NOIND }{  numeric }
  \item{ R24_num }{  character }
  \item{ R24_jour }{  character }
  \item{ R24_nombre }{  numeric }
  \item{ sem_we }{  character }
  \item{ sem_we_ferie }{  character }
  \item{ occ_ouinon }{  character }
  \item{ occ_hdeb }{  hms,difftime }
  \item{ occ_type }{  character }
  \item{ occ_lieu }{  character }
  \item{ occ_lieu_autre_libelle }{  character }
  \item{ occ_lieu_dom_hors_dom }{  character }
}}
\source{
Inca3
}
\usage{
occasions_decode
}
\description{
Data inca3.
}
\keyword{datasets}
