% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_correspondance.R
\docType{data}
\name{table_correspondance}
\alias{table_correspondance}
\title{table_correspondance}
\format{A data frame with 1241 rows and 3 variables:
\describe{
  \item{ format }{  character }
  \item{ code }{  character }
  \item{ libelle }{  character }
}}
\source{
Inca3
}
\usage{
table_correspondance
}
\description{
Data inca3.
}
\keyword{datasets}
