% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inca3-package.R
\docType{package}
\name{inca3-package}
\alias{inca3}
\alias{inca3-package}
\title{inca3: Données de l'étude INCA 3}
\description{
Données de consommations et habitudes alimentaires
    de l'étude INCA 3.Carine Dubuisson, Ariane Dufour, Sandrine Carrillo,
    Peggy Drouillet-Pinard, Sabrina Havard, Jean-Luc Volatier (2019). The
    Third French Individual and National Food Consumption (INCA3) Survey
    2014-2015:method, design and participation rate in the framework of a
    European harmonization process. Public Health Nutrition: 22(4),
    584–600. (doi:10.1017/S1368980018002896).
}
\author{
\strong{Maintainer}: Vincent Guyader \email{vincent@thinkr.fr} (\href{https://orcid.org/0000-0003-0671-9270}{ORCID})

Authors:
\itemize{
  \item Cervan Girard \email{cervan@thinkr.fr}
}

Other contributors:
\itemize{
  \item ANSES [copyright holder]
}

}
\keyword{internal}
